+++ src/preview/index.ts
@@ -8,7 +8,7 @@ import { store, SwingFileType, SwingManifest } from "../store";
 import {
   byteArrayToString,
   getFileContents,
-  stringToByteArray,
+  stringToByteArray
 } from "../utils";
 import { exportSwingToCodePen, registerCodePenCommands } from "./codepen";
 import { registerSwingCommands } from "./commands";
@@ -16,33 +16,34 @@ import { discoverLanguageProviders } from "./languages/languageProvider";
 import {
   getMarkupContent,
   getMarkupExtensions,
-  MARKUP_BASE_NAME,
+  MARKUP_BASE_NAME
 } from "./languages/markup";
 import {
   getReadmeContent,
   README_BASE_NAME,
-  README_EXTENSIONS,
+  README_EXTENSIONS
 } from "./languages/readme";
 import {
   includesReactFiles,
   includesReactScripts,
   REACT_SCRIPTS,
   SCRIPT_BASE_NAME,
-  SCRIPT_EXTENSIONS,
+  SCRIPT_EXTENSIONS
 } from "./languages/script";
 import {
   getStylesheetContent,
   STYLESHEET_BASE_NAME,
-  STYLESHEET_EXTENSIONS,
+  STYLESHEET_EXTENSIONS
 } from "./languages/stylesheet";
 import { createLayoutManager } from "./layoutManager";
 import { getCDNJSLibraries } from "./libraries/cdnjs";
+import { ProxyFileSystemProvider, registerProxyFileSystemProvider } from "./proxyFileSystemProvider";
 import {
   endCurrentTour,
   isCodeTourInstalled,
   registerTourCommands,
   startTourFromUri,
-  TOUR_FILE,
+  TOUR_FILE
 } from "./tour";
 import { registerTutorialModule } from "./tutorials";
 import { storage, TUTORIAL_KEY } from "./tutorials/storage";
@@ -354,7 +355,7 @@ export async function openSwing(uri: Uri) {
     {
       enableScripts: true,
       enableCommandUris: true,
-      localResourceRoots: [uri, currentUri],
+      localResourceRoots: [uri, currentUri, Uri.parse(`${ProxyFileSystemProvider.SCHEME}:/`)],
     }
   );
 
@@ -615,7 +616,8 @@ export function registerPreviewModule(
   registerSwingCommands(context);
   registerTourCommands(context);
   registerCodePenCommands(context);
-
+  registerProxyFileSystemProvider();
+  
   getCDNJSLibraries();
   discoverLanguageProviders();