+++ src/preview/proxyFileSystemProvider.ts
@@ -1,29 +1,52 @@
+import * as path from "path";
+import { URLSearchParams } from "url";
 import * as vscode from "vscode";
+import { byteArrayToString, stringToByteArray } from "../utils";
+import { compileComponent } from "./languages/components/svelte";
+import { processImports } from "./libraries/skypack";
 
-const FS_SCHEME = "codeswing-proxy";
 export class ProxyFileSystemProvider implements vscode.FileSystemProvider {
   private _emitter = new vscode.EventEmitter<vscode.FileChangeEvent[]>();
   readonly onDidChangeFile: vscode.Event<vscode.FileChangeEvent[]> = this
     ._emitter.event;
+
+  static SCHEME = "codeswing-proxy";
 
   stat(uri: vscode.Uri): vscode.FileStat {
     return {
       type: vscode.FileType.File,
       ctime: 0,
       mtime: 0,
-      size: 100000,
+      size: 0,
     };
   }
 
   async readFile(uri: vscode.Uri): Promise<Uint8Array> {
-    const proxyUri = vscode.Uri.parse(decodeURIComponent(uri.path.substr(1)));
-    const file = await vscode.workspace.fs.readFile(proxyUri);
-    return Buffer.from(file.buffer);
+    let proxyUri = vscode.Uri.parse(decodeURIComponent(uri.path.substr(1)));
+    
+    const extension = path.extname(uri.path);
+    if (extension === ".js") {
+      let type;
+      if (uri.query) {
+        const query = new URLSearchParams(uri.query);
+        type = query.get("type");
+        proxyUri = proxyUri.with({ path: proxyUri.path.replace(".js", `.${type}`), query: "" });
+      }
+
+      let contents = byteArrayToString(await vscode.workspace.fs.readFile(proxyUri));
+      if (type === "svelte") {
+        [contents] = await compileComponent(contents);
+      }
+
+      return stringToByteArray(processImports(contents))
+    } else {
+      return vscode.workspace.fs.readFile(proxyUri);
+    }
   }
 
   public static getProxyUri(uri: vscode.Uri) {
     return vscode.Uri.parse(
-      `${FS_SCHEME}:/${encodeURIComponent(uri.toString())}`
+      `${ProxyFileSystemProvider.SCHEME}:/${encodeURIComponent(uri.toString())}`
     );
   }
 
@@ -62,7 +85,9 @@ export class ProxyFileSystemProvider implements vscode.FileSystemProvider {
 
 export function registerProxyFileSystemProvider() {
   vscode.workspace.registerFileSystemProvider(
-    FS_SCHEME,
-    new ProxyFileSystemProvider()
+    ProxyFileSystemProvider.SCHEME,
+    new ProxyFileSystemProvider(), {
+      isReadonly: true
+    }
   );
 }